+++ src/cli/javascript/http_filter.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 
 use futures_util::Future;
 
-use super::channel::{Message, MessageContent};
+use super::channel_mv8::{Message, MessageContent};
 use crate::http::Response;
 use crate::{HttpIO, WorkerIO};
